[BugFix]Fix the bug of setting the default value for the float type when creating table (#42314)

+++ fe/fe-core/src/test/java/com/starrocks/sql/analyzer/AnalyzeDecimalV3Test.java
@@ -893,7 +893,7 @@ public class AnalyzeDecimalV3Test {
             Assert.assertEquals(items.get(0).getType(), ScalarType.DOUBLE);
             Assert.assertEquals(items.get(1).getType(),
                     ScalarType.createDecimalV3Type(PrimitiveType.DECIMAL128, 38, 0));
-            Assert.assertEquals(items.get(2).getType(), ScalarType.DOUBLE);
+            Assert.assertEquals(items.get(2).getType(), ScalarType.FLOAT);
             Assert.assertEquals(items.get(3).getType(), ScalarType.createDecimalV3NarrowestType(3, 2));
             Assert.assertEquals(items.get(4).getType(), ScalarType.createDecimalV3TypeForZero(1));
         }
